terminada la funcionalidad de instanciar los objetos a partir del fichero

+++ src/pagos.java
@@ -1,10 +1,10 @@
 public class pagos {
 
     String idCompra;
-    int cuantia;
+    Double cuantia;
     String fecha;
     
-    public pagos(String idCompra , int cuantia , String fecha){
+    public pagos(String idCompra , Double cuantia , String fecha){
         this.idCompra = idCompra;
         this.cuantia = cuantia;
         this.fecha = fecha;
